fixing more non integer index errors

+++ t-rax/controller/TemperatureController.py
@@ -328,8 +328,8 @@ class TemperatureController(QtCore.QObject):
         self.widget.graph_mouse_pos_lbl.setText("X: {:8.2f}  Y: {:8.2f}".format(x, y))
 
     def roi_mouse_moved(self, x, y):
-        x = np.floor(x)
-        y = np.floor(y)
+        x = int(np.floor(x))
+        y = int(np.floor(y))
         try:
             self.widget.roi_widget.pos_lbl.setText("X: {:5.0f}  Y: {:5.0f}    Int: {:6.0f}    lambda: {:5.2f} nm".
                                                    format(x, y,
